+++ report-generator/test-data/com.teamscale.report.jacoco/TestClassTwo.java
@@ -1,0 +1,13 @@
+
+public class TestClassTwo {
+
+	public void foo() {
+		System.err.println("testing");
+		System.err.println("testing");
+	}
+	
+	public static void main(String[] args) {
+		new TestClassTwo().foo();
+	}
+
+}